Remove all global state.
