+++ registration/frontend/src/admin/cart/index.ts
@@ -1,0 +1,4 @@
+import { Cart } from "./components/Cart";
+import { CartManager, CartResponse } from "./cart-manager";
+
+export { Cart, CartManager, CartResponse };
