+++ registration/frontend/src/admin/mqtt.ts
@@ -2,8 +2,9 @@ import mqtt from "mqtt";
 import mitt, { Emitter } from "mitt";
 
 import { ApisMqttConfig } from "../entrypoints/admin";
+import { Accessor, createSignal, Setter } from "solid-js";
 
-type MqttTopic =
+export type MqttTopic =
   | "refresh"
   | "open"
   | "notification"
@@ -11,10 +12,109 @@ type MqttTopic =
   | "scan/id"
   | "scan/shc";
 
-const emitter: Emitter<Record<MqttTopic, object | null>> = mitt();
-window["mqttEmitter"] = emitter;
+export type MqttEmitter = Emitter<Record<MqttTopic, object | null>>;
 
-const randomClientId = Math.random().toString(16).substr(2, 8);
+export default class MqttClient {
+  public errorMessage: Accessor<string>;
+  private setErrorMessage: Setter<string>;
+
+  public emitter: Emitter<Record<MqttTopic, object | null>>;
+
+  private client: mqtt.MqttClient;
+  private config: ApisMqttConfig;
+
+  constructor(config: ApisMqttConfig) {
+    this.config = config;
+
+    const [errorMessage, setErrorMessage] = createSignal<string>();
+    this.errorMessage = errorMessage;
+    this.setErrorMessage = setErrorMessage;
+
+    this.emitter = mitt();
+
+    const WILDCARD_TOPIC = this.getPrefixedTopic("#");
+
+    const randomClientId = Math.random().toString(16).substr(2, 8);
+    this.client = mqtt.connect(config.broker, {
+      username: config.auth.user,
+      password: config.auth.token,
+      clientId: `${config.auth.user}-${randomClientId}`,
+      clean: true,
+    });
+
+    this.client.on("connect", () => {
+      this.setErrorMessage(undefined);
+      console.debug(`Subscribing to ${WILDCARD_TOPIC}`);
+      this.client.subscribe(WILDCARD_TOPIC, (err) => {
+        if (err) {
+          console.error(`MQTT subscription failed: ${err}`);
+        } else {
+          this.client.publish(
+            this.getPrefixedTopic("admin_presence"),
+            JSON.stringify(":3")
+          );
+        }
+      });
+    });
+
+    this.client.on("error", (err: Error) => {
+      console.error(`MQTT error: ${err}`);
+      this.setErrorMessage(err.toString());
+    });
+
+    this.client.on("reconnect", () => {
+      console.debug("Reconnecting to MQTT");
+    });
+
+    this.client.on("message", (topic, message) => {
+      let data = message.toString();
+      console.debug("MQTT message", topic, data);
+
+      let strippedTopic: MqttTopic;
+      if (topic.startsWith(config.auth.base_topic)) {
+        strippedTopic = topic.slice(
+          config.auth.base_topic.length + 1
+        ) as MqttTopic;
+      } else {
+        console.warn(`Got topic with unexpected prefix: ${topic}`);
+        return;
+      }
+
+      let payload = null;
+      try {
+        payload = JSON.parse(data);
+      } catch (err) {}
+
+      switch (strippedTopic) {
+        case "notification":
+          if (payload?.["text"]) {
+            sendNotification(payload?.["text"]);
+          }
+          break;
+        case "alert":
+          if (payload?.["text"]) {
+            alert(payload?.["text"]);
+          }
+          break;
+        default:
+          this.emitter.emit(strippedTopic, payload);
+          break;
+      }
+    });
+  }
+
+  public disconnect() {
+    this.client.end();
+  }
+
+  public publishMessage(topic: string, payload: string) {
+    this.client.publish(this.getPrefixedTopic(topic), payload);
+  }
+
+  private getPrefixedTopic(topic: string): string {
+    return `${this.config.auth.base_topic}/${topic}`;
+  }
+}
 
 function sendNotification(message: string) {
   if (Notification.permission === "granted") {
@@ -23,87 +123,3 @@ function sendNotification(message: string) {
     alert(message);
   }
 }
-
-export let publishMessage: (topic: string, message: string) => void | null =
-  null;
-
-export function connectToMqtt(config: ApisMqttConfig) {
-  const mqttErrorMessage = document.getElementById("mqtt-client-error");
-
-  function getTopic(topic: string): string {
-    return `${config.auth.base_topic}/${topic}`;
-  }
-
-  const WILDCARD_TOPIC = getTopic("#");
-
-  const client = mqtt.connect(config.broker, {
-    username: config.auth.user,
-    password: config.auth.token,
-    clientId: `${config.auth.user}-${randomClientId}`,
-    clean: true,
-  });
-
-  publishMessage = (topic, message) => {
-    console.debug(`Publishing to topic ${topic} with message: ${message}`);
-    client.publish(getTopic(topic), message);
-  };
-
-  client.on("connect", () => {
-    mqttErrorMessage?.classList?.add("d-none");
-    console.debug(`Subscribing to ${WILDCARD_TOPIC}`);
-    client.subscribe(WILDCARD_TOPIC, (err) => {
-      if (err) {
-        console.error(`MQTT subscription failed: ${err}`);
-      } else {
-        client.publish(getTopic("admin_presence"), JSON.stringify(":3"));
-      }
-    });
-  });
-
-  client.on("error", (err) => {
-    console.error(`MQTT error: ${err}`);
-    mqttErrorMessage?.classList?.remove("d-none");
-  });
-
-  client.on("reconnect", () => {
-    console.debug("Reconnecting to MQTT");
-  });
-
-  client.on("message", (topic, message) => {
-    let data = message.toString();
-    console.debug("MQTT message", topic, data);
-
-    let strippedTopic: MqttTopic;
-    if (topic.startsWith(config.auth.base_topic)) {
-      strippedTopic = topic.slice(
-        config.auth.base_topic.length + 1
-      ) as MqttTopic;
-    } else {
-      console.warn(`Got topic with unexpected prefix: ${topic}`);
-      return;
-    }
-
-    let payload = null;
-    try {
-      payload = JSON.parse(data);
-    } catch (err) {}
-
-    switch (strippedTopic) {
-      case "notification":
-        if (payload?.["text"]) {
-          sendNotification(payload?.["text"]);
-        }
-        break;
-      case "alert":
-        if (payload?.["text"]) {
-          alert(payload?.["text"]);
-        }
-        break;
-      default:
-        emitter.emit(strippedTopic, payload);
-        break;
-    }
-  });
-}
-
-export default emitter;
