+++ registration/frontend/src/admin/scan/components/CloseButton.tsx
@@ -1,0 +1,7 @@
+import { Component } from "solid-js";
+
+export const CloseButton: Component<{ close(): any }> = (props) => {
+  return (
+    <button class="delete" aria-label="delete" onClick={props.close}></button>
+  );
+};
