+++ registration/frontend/src/admin/scan/components/MismatchedData.tsx
@@ -1,0 +1,18 @@
+import { Component, JSX, Show } from "solid-js";
+
+export const MismatchedData: Component<{
+  matched: boolean;
+  message: string;
+  children: JSX.Element;
+}> = (props) => {
+  return (
+    <Show when={!props.matched} fallback={props.children}>
+      <span class="icon-text has-text-warning" title={props.message}>
+        <span class="icon">
+          <i class="fa-solid fa-triangle-exclamation"></i>
+        </span>
+        {props.children}
+      </span>
+    </Show>
+  );
+};
